허밋공격,바닥

+++ Game2/Hermit.cpp
@@ -327,8 +327,8 @@ void Hermit::firebullet()
 		gravity[i] = Vector2(0,0);
 	}
 	
-	firedir[0] = Vector2(RANDOM->Float(50, 80),40);
-	firedir[1] = Vector2(RANDOM->Float(30, 70),40);
+	firedir[0] = Vector2(RANDOM->Float(60, 90),40);
+	firedir[1] = Vector2(RANDOM->Float(30, 60),40);
 
 	bulletcol[0]->SetWorldPos(col->GetWorldPos() + Vector2(40, 340));
 	bulletcol[1]->SetWorldPos(col->GetWorldPos() + Vector2(-90, 360));
@@ -346,8 +346,8 @@ void Hermit::firecannon()
 	}
 	bullet[2]->ChangeAnim(ANIMSTATE::ONCE, 0.09f);
 
-	firedir[0] = Vector2(RANDOM->Float(50, 80), 30);
-	firedir[1] = Vector2(RANDOM->Float(30, 70), 30);
+	firedir[0] = Vector2(RANDOM->Float(60, 90), 30);
+	firedir[1] = Vector2(RANDOM->Float(30, 60), 30);
 	firedir[2] = Vector2(RANDOM->Float(60, 90), 30);
 
 	bulletcol[0]->SetWorldPos(col->GetWorldPos() + Vector2(60, 360));
@@ -355,4 +355,6 @@ void Hermit::firecannon()
 	bulletcol[2]->SetWorldPos(col->GetWorldPos() + Vector2(40, 260));
 
 }
+
+
